best result so far

+++ dataclassutils.py
@@ -237,8 +237,8 @@ class BatchedBuffer:
         rollout_steps = settings.rollout_steps +1
     
         return cls(
-            jnp.zeros((num_envs, rollout_steps, settings.network_settings.obs_size), dtype=jnp.float16),
-            jnp.zeros((num_envs, rollout_steps, settings.network_settings.action_size), dtype=jnp.float16),
+            jnp.zeros((num_envs, rollout_steps, settings.network_settings.obs_size), dtype=jnp.float32),
+            jnp.zeros((num_envs, rollout_steps, settings.network_settings.action_size), dtype=jnp.float32),
             jnp.zeros((num_envs, rollout_steps), dtype=jnp.float32),
             jnp.zeros((num_envs, rollout_steps), dtype=jnp.float32),
             jnp.zeros((num_envs,), dtype=jnp.uint16),
